deployment/requirements.txt
# Core API dependencies
fastapi
uvicorn[standard]
python-multipart
pydantic
paho-mqtt

# Core inference dependencies
numpy>=1.19.0
opencv-python>=4.5.0
pillow